Added another number validation, see description
If the provided RGB numbers are out of range (not within 0-255) the script will set the min/max within this range, but the filename will have the wrong numbers. Because of that this additional validation is added.

Tanks to the user, who generated the following files:
noise_r777-g97777-b7777-(...).png
noise_r-9223372036854775808-(...).png